requirements.txt
discord.py
beautifulsoup4
python-dotenv
playwright